refactor: modularized PAP
