+++ src/app/(protected)/dashboard/_components/UserDashboard/page.tsx
@@ -8,7 +8,7 @@ import { httpGet$GetPAP } from "@/modules/commands/GetPAP/fetcher"
 import { UpdatePAPFetcher$Params } from "@/modules/commands/UpdatePAP/typing"
 import { httpPut$UpdatePAP } from "@/modules/commands/UpdatePAP/fetcher"
 import { Allergen, Allergy, DiscoveryMethod, Symptom } from "@/modules/business-types"
-import { PersonalAllergyProfile } from "./_components/personal-allergy-profile"
+import { PersonalAllergyProfile } from "./_components/PersonalAllergyProfile/page"
 import { httpGet$GetSymptoms } from "@/modules/commands/GetSymptoms/fetcher"
 import { httpGet$GetAllergens } from "@/modules/commands/GetAllergens/fetcher"
 import { httpGet$GetAllergies } from "@/modules/commands/GetAllergies/fetcher"
