+++ src/components/symptom-detail-modal.tsx
